reverted files committed by a mistake

+++ far2l/src/console/interf.hpp
@@ -109,7 +109,7 @@ void ComposeAndSetColor(WORD BaseColor, const FarTrueColorForeAndBack *TFB, bool
 void SetRealColor(DWORD64 wAttributes, bool ApplyToConsole = false);
 DWORD64 GetRealColor();
 void ClearScreen(int Color);
-int GetColor();
+DWORD64 GetColor();
 
 void Box(int x1, int y1, int x2, int y2, int Color, int Type);
 void ScrollBar(int X1, int Y1, int Length, unsigned int Current, unsigned int Total);
